feat: enhance port management with HTTP URL generation and link support - Added external link for HTTP ports in PortInput component, allowing users to visit the generated URL directly. - Refactored PermissionSelector component for improved readability and functionality. - Deleted obsolete RoleInfo component and replaced it with a new structure for role management, including separate edit and user listing components. - Introduced EditPermissions and UsersAssignedToRole components for better separation of concerns in role management.
